+++ common/fractals.c
@@ -60,8 +60,8 @@ _LCMPLX lcoefficient,lold,lnew,lparm, linit,ltmp,ltmp2,lparm2;
 long ltempsqrx,ltempsqry;
 int maxcolor;
 int root, degree,basin;
-double floatmin,floatmax;
-double roverd, d1overd, threshold;
+LDBL floatmin,floatmax;
+LDBL roverd, d1overd, threshold;
 _CMPLX tmp2;
 _CMPLX coefficient;
 _CMPLX  staticroots[16]; /* roots array for degree 16 or less */
@@ -85,7 +85,7 @@ int     bitshiftless1;                  /* bit shift less 1 */
 #define pMPsqr(z) (*pMPmul((z),(z)))
 #define MPdistance(z1,z2)  (*pMPadd(pMPsqr(*pMPsub((z1).x,(z2).x)),pMPsqr(*pMPsub((z1).y,(z2).y))))
 
-double twopi = PI*2.0;
+LDBL twopi = PI*2.0;
 int c_exp;
 
 
@@ -98,17 +98,17 @@ _LCMPLX *longparm; /* used here and in jb.c */
 /*              These variables are external for speed's sake only      */
 /* -------------------------------------------------------------------- */
 
-double sinx,cosx;
-double siny,cosy;
-double tmpexp;
-double tempsqrx,tempsqry;
+LDBL sinx,cosx;
+LDBL siny,cosy;
+LDBL tmpexp;
+LDBL tempsqrx,tempsqry;
 
-double foldxinitx,foldyinity,foldxinity,foldyinitx;
+LDBL foldxinitx,foldyinity,foldxinity,foldyinitx;
 long oldxinitx,oldyinity,oldxinity,oldyinitx;
 long longtmp;
 
 /* These are for quaternions */
-double qc,qci,qcj,qck;
+LDBL qc,qci,qcj,qck;
 
 /* temporary variables for trig use */
 long lcosx, lsinx;
@@ -206,11 +206,11 @@ int near fpANDbailout(void)
 
 int near fpMANHbailout(void)
 {
-   double manhmag;
+   LDBL manhmag;
    tempsqrx=sqr(new.x);
    tempsqry=sqr(new.y);
    magnitude = tempsqrx + tempsqry;
-   manhmag = fabs(new.x) + fabs(new.y);
+   manhmag = fabsl(new.x) + fabsl(new.y);
    if((manhmag * manhmag) >= rqlim) return(1);
    old = new;
    return(0);
@@ -218,7 +218,7 @@ int near fpMANHbailout(void)
 
 int near fpMANRbailout(void)
 {
-   double manrmag;
+   LDBL manrmag;
    tempsqrx=sqr(new.x);
    tempsqry=sqr(new.y);
    magnitude = tempsqrx + tempsqry;
@@ -229,7 +229,7 @@ int near fpMANRbailout(void)
 }
 
 #define FLOATTRIGBAILOUT()  \
-   if (fabs(old.y) >= rqlim2) return(1);
+   if (fabsl(old.y) >= rqlim2) return(1);
 
 #define LONGTRIGBAILOUT()  \
    if(labs(lold.y) >= llimit2) { return(1);}
@@ -239,10 +239,10 @@ int near fpMANRbailout(void)
         { return(1);}
 
 #define FLOATXYTRIGBAILOUT()  \
-   if (fabs(old.x) >= rqlim2 || fabs(old.y) >= rqlim2) return(1);
+   if (fabsl(old.x) >= rqlim2 || fabsl(old.y) >= rqlim2) return(1);
 
 #define FLOATHTRIGBAILOUT()  \
-   if (fabs(old.x) >= rqlim2) return(1);
+   if (fabsl(old.x) >= rqlim2) return(1);
 
 #define LONGHTRIGBAILOUT()  \
    if(labs(lold.x) >= llimit2) { return(1);}
@@ -251,12 +251,12 @@ int near fpMANRbailout(void)
       if(labs((X)) > l16triglim) { return(1);}
 
 #define OLD_FLOATEXPBAILOUT()  \
-   if (fabs(old.y) >= 1.0e8) return(1);\
-   if (fabs(old.x) >= 6.4e2) return(1);
+   if (fabsl(old.y) >= 1.0e8) return(1);\
+   if (fabsl(old.x) >= 6.4e2) return(1);
 
 #define FLOATEXPBAILOUT()  \
-   if (fabs(old.y) >= 1.0e3) return(1);\
-   if (fabs(old.x) >= 8) return(1);
+   if (fabsl(old.y) >= 1.0e3) return(1);\
+   if (fabsl(old.x) >= 8) return(1);
 
 #define LONGEXPBAILOUT()  \
    if (labs(lold.y) >= (1000L<<bitshift)) return(1);\
@@ -286,7 +286,7 @@ int near fpMANRbailout(void)
 
 static int near Halleybailout(void)
 {
-   if ( fabs(modulus(new)-modulus(old)) < parm2.x)
+   if ( fabsl(modulus(new)-modulus(old)) < parm2.x)
       return(1);
    old = new;
    return(0);
@@ -342,7 +342,7 @@ int asmfpMANRbailout(void) { return 0;}
 /* -------------------------------------------------------------------- */
 /*              Fractal (once per iteration) routines                   */
 /* -------------------------------------------------------------------- */
-static double xt, yt, t2;
+static LDBL xt, yt, t2;
 
 /* Raise complex number (base) to the (exp) power, storing the result
 ** in complex (result).
@@ -548,8 +548,8 @@ complex_mult(_CMPLX arg1,_CMPLX arg2,_CMPLX *pz)
 int
 complex_div(_CMPLX numerator,_CMPLX denominator,_CMPLX *pout)
 {
-   double mod;
-   if((mod = modulus(denominator)) < FLT_MIN)
+   LDBL mod;
+   if((mod = modulus(denominator)) < LDBL_MIN)
       return(1);
    conjugate(&denominator);
    complex_mult(numerator,denominator,pout);
@@ -841,7 +841,7 @@ LambdaexponentFractal(void)
    FPUsincos  (&old.y,&siny,&cosy);
 
    if (old.x >= rqlim && cosy >= 0.0) return(1);
-   tmpexp = exp(old.x);
+   tmpexp = expl(old.x);
    tmp.x = tmpexp*cosy;
    tmp.y = tmpexp*siny;
 
@@ -884,8 +884,8 @@ FloatTrigPlusExponentFractal(void)
    /* another Scientific American biomorph type */
    /* z(n+1) = e**z(n) + trig(z(n)) + C */
 
-   if (fabs(old.x) >= 6.4e2) return(1); /* DOMAIN errors */
-   tmpexp = exp(old.x);
+   if (fabsl(old.x) >= 6.4e2) return(1); /* DOMAIN errors */
+   tmpexp = expl(old.x);
    FPUsincos  (&old.y,&siny,&cosy);
    CMPLXtrig0(old,new);
 
@@ -976,12 +976,12 @@ UnityFractal(void)
 int
 UnityfpFractal(void)
 {
-double XXOne;
+LDBL XXOne;
    /* brought to you by Mark Peterson - you won't find this in any fractal
       books unless they saw it here first - Mark invented it! */
 
    XXOne = sqr(old.x) + sqr(old.y);
-   if((XXOne > 2.0) || (fabs(XXOne - 1.0) < ddelmin))
+   if((XXOne > 2.0) || (fabsl(XXOne - 1.0) < ddelmin))
       return(1);
    old.y = (2.0 - XXOne)* old.x;
    old.x = (2.0 - XXOne)* old.y;
@@ -1058,10 +1058,10 @@ longCmplxZpowerFractal(void)
 #ifndef XFRACT
    _CMPLX x, y;
 
-   x.x = (double)lold.x / fudge;
-   x.y = (double)lold.y / fudge;
-   y.x = (double)lparm2.x / fudge;
-   y.y = (double)lparm2.y / fudge;
+   x.x = (LDBL)lold.x / fudge;
+   x.y = (LDBL)lold.y / fudge;
+   y.x = (LDBL)lparm2.x / fudge;
+   y.y = (LDBL)lparm2.y / fudge;
    x = ComplexPower(x, y);
    if(fabs(x.x) < fgLimit && fabs(x.y) < fgLimit) {
       lnew.x = (long)(x.x * fudge);
@@ -1287,7 +1287,7 @@ PopcornFractal(void)
    tempsqrx = sqr(new.x);
    tempsqry = sqr(new.y);
    if((magnitude = tempsqrx + tempsqry) >= rqlim
-     || fabs(new.x) > rqlim2 || fabs(new.y) > rqlim2 )
+     || fabsl(new.x) > rqlim2 || fabsl(new.y) > rqlim2 )
            return(1);
    old = new;
    return(0);
@@ -1410,7 +1410,7 @@ PopcornFractalFn(void)
    tempsqrx = sqr(new.x);
    tempsqry = sqr(new.y);
    if((magnitude = tempsqrx + tempsqry) >= rqlim
-     || fabs(new.x) > rqlim2 || fabs(new.y) > rqlim2 )
+     || fabsl(new.x) > rqlim2 || fabsl(new.y) > rqlim2 )
            return(1);
    old = new;
    return(0);
@@ -2294,7 +2294,7 @@ int
 Magnet1Fractal(void)    /*    Z = ((Z**2 + C - 1)/(2Z + C - 2))**2    */
   {                   /*  In "Beauty of Fractals", code by Kev Allen. */
     _CMPLX top, bot, tmp;
-    double div;
+    LDBL div;
 
     top.x = tempsqrx - tempsqry + floatparm->x - 1; /* top = Z**2+C-1 */
     top.y = old.x * old.y;
@@ -2320,7 +2320,7 @@ Magnet2Fractal(void)  /* Z = ((Z**3 + 3(C-1)Z + (C-1)(C-2)  ) /      */
                     /*       (3Z**2 + 3(C-2)Z + (C-1)(C-2)+1) )**2  */
   {                 /*   In "Beauty of Fractals", code by Kev Allen.  */
     _CMPLX top, bot, tmp;
-    double div;
+    LDBL div;
 
     top.x = old.x * (tempsqrx-tempsqry-tempsqry-tempsqry + T_Cm1.x)
           - old.y * T_Cm1.y + T_Cm1Cm2.x;
@@ -3217,7 +3217,7 @@ int mandphoenix_per_pixel(void)
 int
 QuaternionFPFractal(void)
 {
-   double a0,a1,a2,a3,n0,n1,n2,n3;
+   LDBL a0,a1,a2,a3,n0,n1,n2,n3;
    a0 = old.x;
    a1 = old.y;
    a2 = floatparm->x;
@@ -3272,7 +3272,7 @@ HyperComplexFPFractal(void)
 int
 VLfpFractal(void) /* Beauty of Fractals pp. 125 - 127 */
 {
-   double a, b, ab, half, u, w, xy;
+   LDBL a, b, ab, half, u, w, xy;
 
    half = param[0] / 2.0;
    xy = old.x * old.y;
@@ -3290,7 +3290,7 @@ int
 EscherfpFractal(void) /* Science of Fractal Images pp. 185, 187 */
 {
    _CMPLX oldtest, newtest, testsqr;
-   double testsize = 0.0;
+   LDBL testsize = 0.0;
    long testiter = 0;
 
    new.x = tempsqrx - tempsqry; /* standard Julia with C == (0.0, 0.0i) */
@@ -3433,7 +3433,7 @@ MandelbrotMix4fpFractal(void) /* from formula by Jim Muth */
 #undef K
 #undef L
 
-double b_const;
+LDBL b_const;
 
 int DivideBrot5Setup(void)
 {
